--- Server/model/accounts.js
@@ -1,13 +1,0 @@
-const Accounts = [
-  {
-    id: 1,
-    accountNumber: 1010101010,
-    createdOn: 'Thu Mar 28 2019 16:27:52 GMT+0100 (West Africa Standard Time)',
-    owner: 2,
-    type: 'savings',
-    status: 'active',
-    balance: 1000.00,
-  },
-];
-
-export default Accounts;
