--- Server/model/transaction.js
@@ -1,14 +1,0 @@
-const transactions = [
-  {
-    id: 1,
-    createdOn: 'Mon Apr 01 2019 14:37:21 GMT+0100 (West Africa Standard Time)',
-    type: 'credit',
-    accountNumber: 1010101010,
-    cashier: 1,
-    amount: 1000.00,
-    oldBalance: 1500.00,
-    newBalance: 2500.00,
-  },
-];
-
-export default transactions;
