--- Server/model/userdata.js
@@ -1,31 +1,0 @@
-const Users = [
-  {
-    id: 1,
-    firstName: 'Jude',
-    lastName: 'violet',
-    email: 'jjviolet@gmail.com',
-    hashPassword: '$2a$10$ClJ.aqpzua0fOAIkpcVWven/sPq3OjbZsFCCzBZh2pYntY8gVB8aK',
-    type: 'staff',
-    isAdmin: true,
-  },
-  {
-    id: 2,
-    firstName: 'Amos',
-    lastName: 'Oruroghene',
-    email: 'amoslv@gmail.com',
-    hashPassword: '$2a$10$PQRN6c5NsnGoFcU4mvvgGe67l0W89VvkbduMbij219tXwaJohyZAG',
-    type: 'client',
-    isAdmin: false,
-  },
-  {
-    id: 3,
-    firstName: 'Staff',
-    lastName: 'Staffiene',
-    email: 'staff@gmail.com',
-    hashPassword: '$2a$10$lf1y043YOpqhh5y4EeDYPeVb7PjjsJj4n9FBb5CAUFSOoJGWpAxbe',
-    type: 'staff',
-    isAdmin: false,
-  },
-];
-
-export default Users;
