--- Server/test/testdata/accounts.js
@@ -1,12 +1,0 @@
-const accounts = [
-  {
-    type: 'savings',
-    openingBalance: 1000.00,
-  },
-  {
-    type: 'sav',
-    openingBalance: 'yhyh',
-  },
-];
-
-export default accounts;
