--- Server/test/testdata/user.js
@@ -1,50 +1,0 @@
-const Users = [
-  {
-    firstName: 'James',
-    lastName: 'Chidima',
-    email: 'jchidima@gmail.com',
-    password: 'jude',
-  },
-  {
-    firstName: 'Amos',
-    lastName: 'Oruroghene',
-    email: 'amoslv@gmail.com',
-    password: 'amos',
-  },
-  {
-    firstName: 'Amos',
-    lastName: 'Oruroghene',
-    email: 'amoslvgmailcom',
-    password: 'amos',
-  },
-  {
-    firstName: '1234',
-    lastName: 'Oruroghene',
-    email: 'amoslvw@gmailcom',
-    password: 'amos',
-  },
-  {
-    firstName: 'James',
-    lastName: 'Chidima',
-    email: 'jchidimanonso@gmail.com',
-    password: 'judes',
-  },
-  {
-    firstName: 'James',
-    lastName: 'Chidima',
-    email: 'jchidimanonso@gmail.com',
-    password: 'judeolols',
-  },
-  {
-    firstName: 'Jude',
-    lastName: 'violet',
-    email: 'jjviolet@gmail.com',
-    password: 'jude',
-  },
-  {
-    email: 'staff@gmail.com',
-    password: 'staff',
-  },
-];
-
-export default Users;
