+++ Server_db/router/router.js
@@ -14,7 +14,10 @@ router.post('/auth/signup', Validate.Signup, User.signup);
 router.post('/auth/signin', Validate.Signin, User.signin);
 
 router.post('/accounts', Authenticator.user, Validate.createAccount, Controller.creatAccount);
+router.post('/transactions/:accountNumber/debit', Authenticator.user, Validate.transaction, Controller.transactions);
+router.post('/transactions/:accountNumber/credit', Authenticator.user, Validate.transaction, Controller.transactions);
 
 router.patch('/account/:accountNumber', Authenticator.user, Validate.setAccount, Controller.setAccount);
+
 router.delete('/accounts/:accountNumber', Authenticator.user, Validate.accountNumber, Controller.deleteAccount);
 export default router;
